form show all review

+++ win-project-2/DataClass/NguoiTho.cs
@@ -14,11 +14,11 @@ namespace win_project_2
         public string Skills { get; set; }
         public string JobLocation { get; set; }
         public string Study { get; set; }
-        public int rate { get; set; }
+        public float rate { get; set; }
         public string DonePostIds { get; set; }
 
         // Constructor với tất cả các thuộc tính, kể cả thuộc tính từ lớp cơ sở UserInfo
-        public NguoiTho(string donePostIds, string name, string dateOfBirth, string phoneNumber, string avatarUrl, string email, string address, bool isTho, string id, string jobName, string description, double price, string skills, string jobLocation, string study, int rate)
+        public NguoiTho(string donePostIds, string name, string dateOfBirth, string phoneNumber, string avatarUrl, string email, string address, bool isTho, string id, string jobName, string description, double price, string skills, string jobLocation, string study, float rate)
             : base(name, dateOfBirth, phoneNumber, avatarUrl, email, address, isTho, id)
         {
             JobName = jobName;
